requirements.txt
django
djangorestframework
djoser
djangorestframework-simplejwt
drf-yasg
